requirements.txt
pysnmp>=4.9.0
pyyaml>=6.0
flask>=2.0.0
python-dotenv>=1.0.0  # For environment variable management
requests>=2.31.0  # For HTTP requests if needed
gunicorn>=21.2.0  # Production WSGI server
